# apps/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "API (apps/api)",
    "version": "1.0.0",
    "description": "OpenAPI spec for apps/api"
  },
  "servers": [
    {
      "url": "https://osc2zjnl43.execute-api.us-east-1.amazonaws.com"
    }
  ],
  "paths": {
...
    "/auth/register": {
      "post": {
        "summary": "Register a new user (USER role only)",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "409": {
            "description": "Email already exists"
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Login with email and password",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Logged in successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Invalid credentials"
          }
        }
      }
    },
    "/auth/google": {
      "post": {
        "summary": "Authenticate with Google OAuth profile",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "googleId",
                  "email",
                  "name"
                ],
                "properties": {
                  "googleId": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "name": {
                    "type": "string"
                  },
                  "image": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authenticated via Google",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          }
        }
      }
    },
    "/auth/me": {
      "get": {
        "summary": "Get current authenticated user",
        "tags": [
          "Auth"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Current user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthUser"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          }
        }
      },
      "patch": {
        "summary": "Update current user profile or password",
        "tags": [
          "Auth"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMeInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthUser"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Not authenticated or wrong old password"
          },
          "409": {
            "description": "Email already in use"
          }
        }
      },
      "delete": {
        "summary": "Delete the current authenticated user account",
        "tags": [
          "Auth"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "Account deleted successfully"
          },
          "401": {
            "description": "Not authenticated"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuthUser": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string",
            "nullable": true
          },
          "role": {
            "type": "string",
            "enum": [
              "USER",
              "ADMIN"
            ]
          },
          "image": {
            "type": "string",
            "nullable": true
          },
          "isGoogleAuth": {
            "type": "boolean",
            "description": "True if the user authenticated via Google (google_id is set)"
          },
          "last_login": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "AuthResponse": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/AuthUser"
          },
          "token": {
            "type": "string",
            "description": "JWT access token"
          }
        }
      },
      "RegisterInput": {
        "type": "object",
        "required": [
          "email",
          "password",
          "name",
          "phone"
        ],
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string",
            "minLength": 8
          },
          "name": {
            "type": "string"
          },
          "phone": {
            "type": "string",
            "minLength": 7,
            "example": "+972-54-123-4567"
          }
        }
      },
      "LoginInput": {
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "UpdateMeInput": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "phone": {
            "type": "string"
          },
          "oldPassword": {
            "type": "string",
            "description": "Required when changing password and account has an existing password"
          },
          "newPassword": {
            "type": "string",
            "minLength": 8
          }
        }
      },
...
    }
  },
...
}